dub to float

+++ parallel.cu
@@ -18,7 +18,7 @@
 
 using namespace std;
 
-__global__ void normalize(double * d_input, double *d_max, double *d_min, unsigned int numAttributes, 
+__global__ void normalize(float * d_input, float *d_max, float *d_min, unsigned int numAttributes, 
     unsigned int numElems) {
     
     unsigned int tid = threadIdx.x + blockIdx.x * blockDim.x;
@@ -29,7 +29,7 @@ __global__ void normalize(double * d_input, double *d_max, double *d_min, unsign
     }
 }
 
-__global__ void findDistance(double *d_inputNormal, double *d_inputSample,  double *d_output, unsigned int numElems) {
+__global__ void findDistance(float *d_inputNormal, float *d_inputSample,  float *d_output, unsigned int numElems) {
     unsigned int tid = threadIdx.x + blockIdx.x * blockDim.x;
     if(tid < numElems) {
         d_output[tid] = (d_inputNormal[tid] - d_inputSample[tid])*(d_inputNormal[tid] - d_inputSample[tid]);
@@ -37,7 +37,7 @@ __global__ void findDistance(double *d_inputNormal, double *d_inputSample,  doub
 
 }
 
-/*__global__ void block_sum(double *input, double *results, size_t n)
+/*__global__ void block_sum(float *input, float *results, size_t n)
 {
     extern __shared__ float sdata[];
     int i = threadIdx.x + blockDim.x*blockIdx.x;
@@ -65,8 +65,8 @@ __global__ void findDistance(double *d_inputNormal, double *d_inputSample,  doub
 }*/
 
 void parse(int* numAttributes, int* numKnownSamples, int* numClass, int *numUnknowns,
-    double ** min, double ** max, double ** knowns, int ** classifications, 
-    double ** unknowns, string** unknownNames)
+    float ** min, float ** max, float ** knowns, int ** classifications, 
+    float ** unknowns, string** unknownNames)
 {
     ifstream myfile(FILE_NAME, ios::in);  // declare and open
     
@@ -79,8 +79,8 @@ void parse(int* numAttributes, int* numKnownSamples, int* numClass, int *numUnkn
     *numUnknowns = numUn;
     
     // Populate all of the mins and maxes
-    *min = (double*) malloc(sizeof(double) * numAttrib);
-    *max = (double*) malloc(sizeof(double) * numAttrib);
+    *min = (float*) malloc(sizeof(float) * numAttrib);
+    *max = (float*) malloc(sizeof(float) * numAttrib);
     for (int i = 0; i < numAttrib; i++) {
         int currentMax, currentMin;
         myfile >> currentMin >> currentMax;
@@ -91,7 +91,7 @@ void parse(int* numAttributes, int* numKnownSamples, int* numClass, int *numUnkn
     
     // Populate the known object types
     *classifications =(int*) malloc(sizeof(int) * numKnownSamp);
-    *knowns = (double*) malloc(sizeof(double) * numKnownSamp * numAttrib);
+    *knowns = (float*) malloc(sizeof(float) * numKnownSamp * numAttrib);
     
     for (int i = 0; i < numKnownSamp; i++) {
         int currentClass;
@@ -99,7 +99,7 @@ void parse(int* numAttributes, int* numKnownSamples, int* numClass, int *numUnkn
         (*classifications)[i] = currentClass;
         
         for (int j = 0; j < numAttrib; j++) {
-            double currentAttrib;
+            float currentAttrib;
             myfile >> currentAttrib;
             (*knowns)[i*numAttrib + j] = currentAttrib;
         }
@@ -107,7 +107,7 @@ void parse(int* numAttributes, int* numKnownSamples, int* numClass, int *numUnkn
     
     // Populate the unknown object types
     *unknownNames = new string[numUn];
-    *unknowns = (double*) malloc(sizeof(double) * numUn * numAttrib);
+    *unknowns = (float*) malloc(sizeof(float) * numUn * numAttrib);
     
     for (int i = 0; i < numUn; i++) {
         string currentName;
@@ -115,7 +115,7 @@ void parse(int* numAttributes, int* numKnownSamples, int* numClass, int *numUnkn
         (*unknownNames)[i] = currentName;
         
         for (int j = 0; j < numAttrib; j++) {
-            double currentAttrib;
+            float currentAttrib;
             myfile >> currentAttrib;
             (*unknowns)[i*numAttrib + j] = currentAttrib;
         }
@@ -135,38 +135,38 @@ int main() {
     int numAttributes, numKnownSamples, numClass, numUnknowns;
     
     // Data that needs to be sent to the device.
-    double *h_min, *h_max;
-    double *h_knowns;
+    float *h_min, *h_max;
+    float *h_knowns;
     int *h_classifications;
-    double *h_unknowns;
+    float *h_unknowns;
     
     // Device data
-    double *d_min, *d_max;
-    double *d_knowns;
+    float *d_min, *d_max;
+    float *d_knowns;
     int *d_classifications;
-    double *d_unknowns;
+    float *d_unknowns;
     
     string *unknownNames;
     
     // Needed for the profiling
     std::clock_t start;
-    double duration;
+    float duration;
     
     parse(&numAttributes, &numKnownSamples, &numClass, &numUnknowns, 
         &h_min, &h_max, &h_knowns, &h_classifications, &h_unknowns, &unknownNames);
     
     // Start mallocing the data to the kernel
-    cudaMalloc(&d_min, sizeof(double) * numAttributes);
-    cudaMalloc(&d_max, sizeof(double) * numAttributes);
-    cudaMalloc(&d_knowns, sizeof(double) * numKnownSamples * numAttributes);
-    cudaMalloc(&d_unknowns, sizeof(double) * numUnknowns * numAttributes);
+    cudaMalloc(&d_min, sizeof(float) * numAttributes);
+    cudaMalloc(&d_max, sizeof(float) * numAttributes);
+    cudaMalloc(&d_knowns, sizeof(float) * numKnownSamples * numAttributes);
+    cudaMalloc(&d_unknowns, sizeof(float) * numUnknowns * numAttributes);
     cudaMalloc(&d_classifications, sizeof(int) * numKnownSamples);
     
     // Copy the data from the host to the kernel
-    cudaMemcpy(d_min, h_min, sizeof(double) * numAttributes, cudaMemcpyHostToDevice);
-    cudaMemcpy(d_max, h_max, sizeof(double) * numAttributes, cudaMemcpyHostToDevice);
-    cudaMemcpy(d_knowns, h_knowns, sizeof(double) * numKnownSamples * numAttributes, cudaMemcpyHostToDevice);
-    cudaMemcpy(d_unknowns, h_unknowns, sizeof(double) * numUnknowns * numAttributes, cudaMemcpyHostToDevice);
+    cudaMemcpy(d_min, h_min, sizeof(float) * numAttributes, cudaMemcpyHostToDevice);
+    cudaMemcpy(d_max, h_max, sizeof(float) * numAttributes, cudaMemcpyHostToDevice);
+    cudaMemcpy(d_knowns, h_knowns, sizeof(float) * numKnownSamples * numAttributes, cudaMemcpyHostToDevice);
+    cudaMemcpy(d_unknowns, h_unknowns, sizeof(float) * numUnknowns * numAttributes, cudaMemcpyHostToDevice);
     cudaMemcpy(d_classifications, h_classifications, sizeof(int) * numKnownSamples, cudaMemcpyHostToDevice);
 
     
@@ -182,14 +182,14 @@ int main() {
     normalize<<<numBlocks, threadsPerBlock>>>(d_unknowns, d_max, d_min, 
         numAttributes, numUnknowns);
     
-    cudaMemcpy(h_unknowns, d_unknowns, sizeof(double) * numUnknowns * numAttributes, cudaMemcpyDeviceToDevice);
+    cudaMemcpy(h_unknowns, d_unknowns, sizeof(float) * numUnknowns * numAttributes, cudaMemcpyDeviceToDevice);
     
     for (int i = 0; i < 5; i++) {
         printf("%f ", h_unknowns[i]); 
     }
     printf("\n");
 
-
-
-    //findDistance<<<numBlocks, threadsPerBlock>>>(d_knowns, 
+    //float * distance;
+    //cudaMalloc(&distance, sizeof(float) * numSomething
+    //findDistance<<<numBlocks, threadsPerBlock>>>(d_knowns, d_unknowns, d_distance,  
 }
